Doxygen documentation
The intent is to generate documentation only for functions that are
designed to be part of the public API and therefore documented with
Doxygen comments. No documentation will be generated for code without
such comments. Note that Doxygen only generates documentation for free
functions if they are in a file with a @file comment.

Update the README to point at the documentation that will be generated
and hosted by GitLab Pages.

The main page generated by Doxygen shall become the index page for the
IntrinsicCV GitLab Pages page. Therefore include a link from that main
page to the coverage page.

Doxygen doesn't handle __VA_OPT__ well so split INTRINSICCV_BINARY_OP
into two more specialised macros.

Add @copydoc comments to generate documentation for similar functions
with minimal copy-paste.

GitLab supports customising the directory from which pages are uploaded
so use a build subdirectory instead of public, to limit how messy the
source folder gets.

Add Doxygen to Dockerfile. Doxygen may use GraphViz to create graphs so
install that as well.

While updating the Dockerfile it was found that it wouldn't build
because LLVM is now at version 19 so fix the LLVM version.