Handle zero sized images
Where it is possible to handle zero sized images in a valid way (e.g.
mapping from one zero sized image to another) then permit such inputs.

For functions that must return some property of the image that will be
undefined if it has no size, require that the image size must be greater
than 1.