Disable cv_hal_minMaxIdx implementation
OpenCV 4.9 has a bug that cv_hal_minMaxIdx is called regardless of the
number of channels in the image even though the interface has no way to
pass the channel count to it.

A fix is planned in OpenCV 4.10 but as yet it is unstable. Therefore
disable the implementation in KleidiCV for now.