Fix SVE2/SME2 path of exp_f32 for some inputs
In case of SVE2/SME2 the order of some multiplications was incorrect
and for inputs resulting almost positive infinte values (like the input
value of 88.7F) the calculation was faulty.

In case of NEON the change is not functional, it just makes easier to
the reader to understand the order of multiplications.