Use __arm_streaming attribute for SME functions
As the library avoids branching as much as possible there is no reason
to share functions in the binary between the SVE2 and SME/SME2 paths.
That kind of sharing is made possible by the __arm_streaming_compatible
attribute.
In addition, functions using SME2 instructions must be attributed with
the __arm_streaming (or __arm_locally_streaming) keyword.