Fix the inline asm in gray_to_rgba and rgb_to_bgr
Memory is loaded inside the inline assembly to control which registers
are used as input in case of tbl instrunctions. (The tbl instruction
requires a continious range of table registers, but the compiler can
choose any register to be passed to the inline assmebly part. If the
passed table registers are not continous code generation fails.)

Force inlining is also added to avoid branches in case of GCC.